Plugable pythonic classes for Murano
One of the key features of Murano is extensibility, and we need to
push this feature even further and give our customers a way to
extend Murano with new functions (e.g. support for F5 BigIP API)
in a drag-n-drop manner.

This spec proposes a solution which will add this extensibility
option.

Change-Id: I63dfecd4740cc8d10ea45a6f36a633b76c6a883d