Support TOSCA definitions for applications
TOSCA is a standard developed under OASIS foundation. It is aimed
to cover definition of complex enterprise applications which might
consist of different loosely coupled components. Components can be
tied by using requirements and capabilities which are part of
TOSCA standard. As TOSCA covers most of the Applications aspects
(excluding application compilation and build) it should be
straightforward to support TOSCA format in murano. TOSCA is
adopted by enterprises so it will allow the openstack ecosystem to
integrate with enterprise IT applications.

Partially Implements: blueprint support-tosca-format

Change-Id: I79d57e838bd7c8ec054194b87d146f505eec6f75