Add network selection element to UI form
Sometime a VM should be placed to existing network rather then to a new
network created during deployment. While our workflows support this,
there is no way for the end user to select this network in the easy way
in the UI. It will be great if there is a special form element which
will pre-populate a list of available networks and provide an easy
option to select desired network for the application.

Spec for blueprint ui-network-selection

Change-Id: I94b4017ec466be09bddc6df01b975bbad63cac3b