Glance Artifact Repository Support
Glance has recently got a new feature - ability to store not only the VM images
but also data assets and their metadata for other OpenStack projects.

This specification defines the usage of this feature in Murano, so Murano may
store its packages in Glance and benefit from all its features.

Change-Id: I26c2d495d4a4178c2794e315a254f93525546a6f