Support for Service Function Chaining in Murano
One of the foundations for NFV enabled clouds is to have Networking
Service Function Chaining which provides an ability to define an ordered
list of network services which to form a “chain” of services.
This could be used by f.e. Telcos to simplify management of their
infrastructure.

Implements blueprint: plugin-networking-sfc
Change-Id: I2fff8f01949a708e0251ad860fbd0a238c9e92fc