Change the naming scheme for MuranoPL plugins
Current naming scheme for classes defined in MuranoPL extension
plugins is not consistent with the naming scheme for regular MuranoPL
classes and packages.

This spec addresses this by changing the scheme and proposing a
fallback logic which will provide a backward compatibility for the
existing MuranoPL applications targeting old plugins.

Change-Id: I05da65257b0002a2ecafe2b680d9ee6495b06474